+++ backend/canisters/user/impl/src/updates/c2c_local_user_index.rs
@@ -2,7 +2,7 @@ use crate::guards::caller_is_local_user_index;
 use crate::{RuntimeState, execute_update, openchat_bot};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
-use types::{Achievement, DiamondMembershipPlanDuration, ReferralStatus, Timestamped};
+use types::{Achievement, DiamondMembershipPlanDuration, IdempotentEnvelope, ReferralStatus, Timestamped};
 use user_canister::c2c_local_user_index::*;
 use user_canister::mark_read::ChannelMessagesRead;
 use user_canister::{LocalUserIndexEvent, UserCanisterEvent};
@@ -10,11 +10,11 @@ use user_canister::{LocalUserIndexEvent, UserCanisterEvent};
 #[update(guard = "caller_is_local_user_index", msgpack = true)]
 #[trace]
 fn c2c_local_user_index(args: Args) -> Response {
-    execute_update(|state| c2c_local_user_index_impl(args, state))
+    execute_update(|state| handle_events(args.events, state))
 }
 
-fn c2c_local_user_index_impl(args: Args, state: &mut RuntimeState) -> Response {
-    for event in args.events {
+pub(crate) fn handle_events(events: Vec<IdempotentEnvelope<LocalUserIndexEvent>>, state: &mut RuntimeState) -> Response {
+    for event in events {
         if state.data.idempotency_checker.check(
             state.data.local_user_index_canister_id,
             event.created_at,